requirements.txt
django
djangorestframework
markdown
django-filter
django-environ
django-extensions
django[bcrypt]
# psycopg2-binary
# gunicorn